+++ CHANGELOG.md
@@ -1,3 +1,13 @@
+#2.0.2
+- New coupon conditions :
+    - Start date
+    - Billing country
+    - Shipping country
+    - Cart contains product
+    - Cart contains product from category
+    - For specific customers
+    
+
 #2.0.1
 - possibility to apply a permanent discount on a customer
 - display estimated shipping on cart page